Backend
The node_modules file is just what needed to be installed on my computer
for node to work on the command line, I don't know if you guys will need
it or not. The views folder is where we can keep the html files, or the
files that the user will actually see. The routes folder has .js files
that the app.js file will reference to.